seperate summarypage from app
